Started calibration function
Logically it should be the same, the idea is to have the motor run a distance press a button that position is saved as position 0. The motor goes in the opposite direction and same idea, but with each movement a count is kept at the end the count is saved as the max distance available.