Calibration sample code
This code is only for one motor the horizontal is all that this currently works for. Line 48 in the code does not work as expected. For some reason when going in the opposite direction the horizontal motor doesn't step "properly". Mostly a bad understanding of how the moveTo function is working.